pydantic-server/pyproject.toml
[project]
name = "pydantic-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai==0.0.52",
    "pydantic-ai-slim[duckduckgo]==0.0.52",
    "arize-phoenix",
    "langchain-community>=0.3.21",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "openinference-instrumentation-openai>=0.1.23",
    "jinja2>=3.1.6",
    "wikipedia>=1.4.0",
    "langchain-google-community>=2.0.7",
    "chromadb-client>=1.0.0",
    "onnxruntime>=1.21.0",
    "chromadbx>=0.1.0",
]
